CPP03/ex00 => checks before correction

+++ CPP03/ex00/includes/ClapTrap.hpp
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ClapTrap.hpp                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: antoine <antoine@student.42.fr>            +#+  +:+       +#+        */
+/*   By: anvincen <anvincen@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/07/24 12:39:16 by anvincen          #+#    #+#             */
-/*   Updated: 2023/10/09 09:54:44 by antoine          ###   ########.fr       */
+/*   Updated: 2023/10/10 17:43:34 by anvincen         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,10 +19,10 @@
 class	ClapTrap
 {
 	private :
-		std::string	_name;
-		size_t		_hitPoint;
-		size_t		_energyPoint;
-		size_t		_attackDamage;
+		std::string		_name;
+		unsigned int	_hitPoint;
+		unsigned int	_energyPoint;
+		unsigned int	_attackDamage;
 
 	public :
 		ClapTrap( void );
